L-Variant/producers/iiot-simulator/requirements.txt
kafka-python
numpy
PyYAML
pandas
faust